requirements.txt
pandas
numpy
scikit-learn
imbalanced-learn
flask
flask-cors
joblib
praat-parselmouth
nolds
scipy
